requirements.txt
asgiref==3.2.10
cached-property==1.5.2
click==7.1.2
Django==3.1.2
django-autocompletefilter==0.0.6
django-choices==1.7.1
django-colorfield==0.1.15
django-cors-headers==3.7.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-webpack-loader==0.7.0
gunicorn==20.0.4
graphql-core==3.1.3
hupper==1.10.2
mypy-extensions==0.4.3
numpy==1.19.4
pathlib==1.0.1
Pint==0.9
psycopg2-binary==2.8.6
Pygments==2.8.1
pyparsing==2.4.0
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
starlette==0.14.2
strawberry-graphql==0.56.0
strawberry-graphql-django==0.1.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
whitenoise==5.0.1
